document-indexer/requirements.txt
pika==1.3.2
retry==0.9.2
redis==4.6.0
pdfplumber==0.10.0
qdrant-client==1.3.1
aiohttp==3.8.4
azure-identity==1.13.0
azure-storage-blob==12.17.0